glwidgets: Fix the 3D viewport model appearing white and buggy on startup. This was caused by overflowing the uint8_t on the class structure and setting the following fields that contain the render modes to zero. Don't cast the write and stop trying to be smart by using smaller types to my own detriment. Bug introduced in 3813f1ed0.

+++ glwidgets.h
@@ -75,7 +75,7 @@ public:
 	void setMaterialError(int newErr);
 
 	QSet<QString> supportedExtensionsList;
-	uint8_t maxSupportedTexAnisoTaps = 0;
+	int32_t maxSupportedTexAnisoTaps = 0;
 
 	void keyPressEvent( QKeyEvent * event );
 	void keyReleaseEvent( QKeyEvent * event );
